+++ google2/CHANGELOG.md
@@ -7,11 +7,12 @@ Changed
 - Bump `fs2-io` to `2.0.1`
 - Add optional `blockderBound` to `GoogleStorageService` constructor
 - Remove `LineBacker` usage
+- Add arguments to `insertBucket` 
 
 Add
 - Add `GoogleDataproc` and `GoogleDataprocInterpreter`
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.6-8bae8e8"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.6-TRAVIS-REPLACE-ME"`
 
 ## 0.5